// frontend_temp/src/types/global.d.ts
export type Bonuses = {
	amount?: number
	reason?: string
	date?: string
	_id?: string
}

export type Deductions = {
	amount?: number
	reason?: string
	date?: string
	_id?: string
}

interface SalaryDetails {
	baseSalary?: number;
	bonuses?: Bonuses[];
	deductions?: Deductions[]
}

interface SemesterFees {
	semester?: number;
	totalFee?: number;
	adjustedFee?: number;
	dueDate?: string;
	status?: string;
	_id?: string;
}

interface feeDetails {
	scholorShip?: number;
	totalFee?: number;
	adjustedFee?: number;
	paidFee?: number;
	dueFee?: number;
	semesterFees?: SemesterFees[]
}
// ...
export interface StudentType {
	_id: string;
	fullName?: string;
	email?: string;
	password?: string;
	contactNumber?: string;
	gender?: string;
	address?: string;
	createdAt?: string;
	updatedAt?: string;
	role?: string;
	dob?: string;
	status?: string;
	program?: string;
	rollNumber?: number;
	sessionYears?: string;
	class?: string;
	section?: string;
	feeDetails?: feeDetails
}
// ...
export interface StudentsFilter {
	limit?: number;
	page?: number;
	rollNumber?: string;
	fullName?: string;
	gender?: string;
	program?: string;
	status?: string;
	classId?: string;
	sectionId?: string;
	sortBy?: string;
	sortType?: string;
}

export interface TeacherType {
	_id: string;
	fullName?: string;
	email?: string;
	password?: string;
	contactNumber?: string;
	gender?: string;
	dob?: string;
	address?: string;
	createdAt?: string;
	updatedAt?: string;
	departments?: string[];
	subjectSpecialization?: string;
	salaryDetails: {
		baseSalary: number;
		bonuses: Bonuses[];
		deductions: Deductions[];
	}
	status?: string;
	totalSalary?: number; // This is for table data
}

export interface DepartmentType {
	_id: string;
	name?: string;
	headOfDept?: any;
}

// for use in api
export interface DepartmentsResponseArray extends ApiResponse {
	data: {
		pagination: any;
		departments: DepartmentType[];
	};
}

export interface DepartmentsFiler {
	limit?: number;
	page?: number;
}

// for use in api
export interface TeachersResponseArray extends ApiResponse {
	data: {
		pagination: any;
		teachers: TeacherType[];
	};
}

export interface TeachersFilter {
	limit?: number;
	page?: number;
	sortBy?: string;
	filterType?: string;
	sortType?: asc | desc;
	salaryFilterType?: string | undefined | null;
	subjectSpecialization?: string;
	status?: string;
	fullName?: string;
	email?: string;
	gender?: string;
	baseSalary?: number | null;
	departments?: string; // single department
}

export interface SingleResponseTeacherType extends ApiResponse {
	data: TeacherType
}


// TIMETABLES..........
export interface TimetableType {
	_id: string;
	teacher?: string;
	section?: string;
	subject: string;
	day: string;
	timeslot: {
		start: string;
		end: string;
	};
	isOngoing: boolean;
}
// ...
export interface StudentAttendanceFilterState {
	limit: number;
	page: number;
	sectionId: string;
}